Add WIP refactored code
This R code is ultimately intended to replace the pipeline's dependence on (proprietary) Mathematica. It handles network construction (from trajectory XML files), analysis, and visualisation, and as such should currently be sufficient to reproduce the key results in the paper. Ongoing additions will take care of the other network statistics.